Refactored

+++ src/Infrastructure/Booking.Auth.Persistence/Repositories/CompanyRepository.cs
@@ -21,13 +21,16 @@ public class CompanyRepository(DataContext context) : BaseRepository<Company>(co
             .AnyAsync(x => x.Id != id && x.Inn == inn, cancellationToken);
     }
 
-    public async Task<List<Company>> GetAllCompaniesAsync(int offset, int limit, CancellationToken cancellationToken = default)
+    public async Task<Tuple<List<Company>, int>> GetAllCompaniesAsync(int offset, int limit, CancellationToken cancellationToken = default)
     {
-        return await Context.Companies.Skip(offset).Take(limit).ToListAsync(cancellationToken);
+        return new Tuple<List<Company>, int>(await Context.Companies.Skip(offset).Take(limit).ToListAsync(cancellationToken), await Context.Companies.CountAsync(cancellationToken));
     }
 
     public async Task DeleteCompanyByIdAsync(Guid id, CancellationToken cancellationToken = default)
     {
-        await Context.Companies.Where(elm => elm.Id == id).ExecuteDeleteAsync(cancellationToken);
+        bool any = await Context.Companies.AnyAsync(elm => elm.Id == id, cancellationToken);
+
+        if (any)
+            await Context.Companies.Where(elm => elm.Id == id).ExecuteDeleteAsync(cancellationToken);
     }
 }
