+++ src/Infrastructure/Booking.Auth.Persistence/Repositories/FilialRepository.cs
@@ -21,13 +21,16 @@ public class FilialRepository(DataContext context) : BaseRepository<Filial>(cont
             .AnyAsync(x => x.Id != id && x.Name == name, cancellationToken);
     }
 
-    public async Task<List<Filial>> GetFilialsListAsync(int offset, int limit, CancellationToken cancellationToken = default)
+    public async Task<Tuple<List<Filial>, int>> GetFilialsListAsync(int offset, int limit, CancellationToken cancellationToken = default)
     {
-        return await Context.Filials.Skip(offset).Take(limit).ToListAsync(cancellationToken);
+        return new Tuple<List<Filial>, int>(await Context.Filials.Skip(offset).Take(limit).ToListAsync(cancellationToken), await Context.Filials.CountAsync(cancellationToken));
     }
 
     public async Task DeleteFilialByIdAsync(Guid id, CancellationToken cancellationToken = default)
     {
-        await Context.Filials.Where(elm => elm.Id == id).ExecuteDeleteAsync(cancellationToken);
+        bool any = await Context.Filials.AnyAsync(elm => elm.Id == id, cancellationToken);
+
+        if(any)
+            await Context.Filials.Where(elm => elm.Id == id).ExecuteDeleteAsync(cancellationToken);
     }
 }
